+++ tapo/src/api/discovery.rs
@@ -1,6 +1,8 @@
 mod aes_discovery_query_generator;
 mod device_discovery;
+mod device_type;
 mod discovery_result;
 
 pub use device_discovery::*;
+pub use device_type::*;
 pub use discovery_result::*;